+++ src/components/form/form-input/index.js
@@ -1,0 +1,3 @@
+import FormInput from "./form-input.vue";
+
+export default { FormInput };