+++ src/components/form/form-label/index.js
@@ -1,0 +1,3 @@
+import FormLabel from "./form-label.vue";
+
+export default { FormLabel };