+++ src/components/icon/icon-danger.cy.js
@@ -1,0 +1,13 @@
+import IconDanger from "./icon-danger.vue";
+import { createMount } from "@cypress/support/mount";
+
+const defaultProps = { class: "stroke-blue-800" };
+const mount = createMount(IconDanger, { props: defaultProps });
+
+describe("icon-danger", () => {
+	it("Renders an icon", () => {
+		mount();
+
+		cy.get("svg").shouldBeVisible();
+	});
+});