+++ src/components/index.js
@@ -1,20 +1,24 @@
-import UiButton from "./ui-button/ui-button.vue";
+import FormInput from "./form/form-input/form-input.vue";
 import LinkTag from "./link-tag/link-tag.vue";
+import UiButton from "./ui-button/ui-button.vue";
 
 import BaseIcon from "./icon/base-icon.vue";
 import IconChevronDown from "./icon/icon-chevron-down.vue";
 import IconChevronLeft from "./icon/icon-chevron-left.vue";
 import IconChevronRight from "./icon/icon-chevron-right.vue";
 import IconChevronUp from "./icon/icon-chevron-up.vue";
+import IconDanger from "./icon/icon-danger.vue";
 import IconExternal from "./icon/icon-external.vue";
 import IconLoading from "./icon/icon-loading.vue";
 
 export default {
 	BaseIcon,
+	FormInput,
 	IconChevronDown,
 	IconChevronLeft,
 	IconChevronRight,
 	IconChevronUp,
+	IconDanger,
 	IconExternal,
 	IconLoading,
 	LinkTag,