+++ src/preview/preview-section.vue
@@ -7,11 +7,3 @@
 		<slot />
 	</div>
 </template>
-
-<script setup>
-import { ref } from "vue";
-
-const props = defineProps({
-
-});
-</script>